update devnet node setup instructions

+++ scripts/devnet-val-setup.sh
@@ -1,37 +1,48 @@
-echo "Install dependencies"
-sudo apt update
-sudo apt install build-essential jq -y
+#!/bin/bash
 
-wget https://dl.google.com/go/go1.15.2.linux-amd64.tar.gz
-tar -xvf go1.15.2.linux-amd64.tar.gz
-sudo mv go /usr/local
+command_exists () {
+    type "$1" &> /dev/null ;
+}
 
-echo "" >> ~/.bashrc
-echo 'export GOPATH=$HOME/go' >> ~/.bashrc
-echo 'export GOROOT=/usr/local/go' >> ~/.bashrc
-echo 'export GOBIN=$GOPATH/bin' >> ~/.bashrc
-echo 'export PATH=$PATH:/usr/local/go/bin:$GOBIN' >> ~/.bashrc
+if command_exists go ; then
+    echo "Golang is already installed"
+else
+  echo "Install dependencies"
+  sudo apt update
+  sudo apt install build-essential jq -y
 
-#source ~/.bashrc
-. ~/.bashrc
+  wget https://dl.google.com/go/go1.15.2.linux-amd64.tar.gz
+  tar -xvf go1.15.2.linux-amd64.tar.gz
+  sudo mv go /usr/local
 
-go version
+  echo "" >> ~/.bashrc
+  echo 'export GOPATH=$HOME/go' >> ~/.bashrc
+  echo 'export GOROOT=/usr/local/go' >> ~/.bashrc
+  echo 'export GOBIN=$GOPATH/bin' >> ~/.bashrc
+  echo 'export PATH=$PATH:/usr/local/go/bin:$GOBIN' >> ~/.bashrc
 
--- Install Regen-ledger and setup the node -----
+  #source ~/.bashrc
+  . ~/.bashrc
+
+  go version
+fi
+
+echo "-- Clear old regen data and install Regen-ledger and setup the node --"
 
 rm -rf ~/.regen
 rm -rf $GOPATH/src/github.com/regen-network/regen-ledger
 
-YOUR_KEY_NAME=mykey
-YOUR_NAME=myname
+YOUR_KEY_NAME=$1
+YOUR_NAME=$2
 DAEMON=regen
 DENOM=utree
-CHAIN_ID=regen-devnet-1
-PERSISTENT_PEERS="a621e6bf1f5981b3e72e059f86cbfc9dc5577fcb@18.220.101.192:26656"
+CHAIN_ID=regen-devnet-2
+PERSISTENT_PEERS="f864b879f59141d0ad3828ee17ea0644bdd10e9b@18.220.101.192:26656"
 
 echo "install regen-ledger:master"
 git clone https://github.com/regen-network/regen-ledger $GOPATH/src/github.com/regen-network/regen-ledger
 cd $GOPATH/src/github.com/regen-network/regen-ledger
+git checkout v0.6.0-alpha2
 make install
 
 echo "Creating keys"
@@ -68,17 +79,11 @@ sudo mv $DAEMON.service /lib/systemd/system/$DAEMON.service
 sudo -S systemctl daemon-reload
 sudo -S systemctl start $DAEMON
 
-
-echo "Your node setup is done. You would need some tokens to start your validator. You can get some tokens from the faucet: https://faucet.devnet.regen.vitwit.com"
-
 echo
-
 echo "Your account address is :"
 $DAEMON keys show $YOUR_KEY_NAME -a
-
+echo "Your node setup is done. You would need some tokens to start your validator. You can get some tokens from the faucet: https://faucet.devnet.regen.vitwit.com"
 echo
 echo
-
 echo "After receiving tokens, you can create your validator by running"
-
 echo "$DAEMON tx staking create-validator --amount 90000000000$DENOM --commission-max-change-rate \"0.1\" --commission-max-rate \"0.20\" --commission-rate \"0.1\" --details \"Some details about yourvalidator\" --from $YOUR_KEY_NAME   --pubkey=\"$($DAEMON tendermint show-validator)\" --moniker $YOUR_NAME --min-self-delegation \"1\" --chain-id $CHAIN_ID --node http://18.220.101.192:26657"
